requirements.txt
anthropic>=0.25.0
chromadb>=0.4.0
sentence-transformers>=2.7.0
rouge-score>=0.1.2
scikit-learn>=1.4.0
nltk>=3.8.0
PyMuPDF>=1.24.0
python-docx>=1.1.0
pandas>=2.2.0
numpy>=1.26.0
tqdm>=4.66.0
python-dotenv>=1.0.0